refac: move date tests to date utils test file
