+++ pkg/utils/dateutils_test.go
@@ -18,6 +18,7 @@
 package utils
 
 import (
+	"encoding/json"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -160,4 +161,54 @@ func Test_SanitizeHistogramInterval(t *testing.T) {
 	assert.Nil(t, err)
 	assert.Equal(t, expected, actual, "expected  %v, actual %v", expected, actual)
 
+}
+
+func Test_ExtractTimeStamp(t *testing.T) {
+	m := make(map[string]interface{})
+
+	// Case 1: timestamp as json.Number
+	m["timestamp"] = json.Number("950823120000")
+	rawJson, _ := json.Marshal(m)
+	tskeyCfg := "timestamp"
+	ts_millis := ExtractTimeStamp(rawJson, &tskeyCfg)
+	assert.Equal(t, uint64(950823120000), ts_millis)
+
+	// Case 2: timestamp as String
+	m["timestamp"] = "950823120000"
+	rawJson, _ = json.Marshal(m)
+	ts_millis = ExtractTimeStamp(rawJson, &tskeyCfg)
+	assert.Equal(t, uint64(950823120000), ts_millis)
+}
+
+func Test_ConvertTimestampToMillis(t *testing.T) {
+	// valid timestamps
+	value := "950823120000"
+	ts_millis, _ := ConvertTimestampToMillis(value)
+	assert.Equal(t, uint64(950823120000), ts_millis)
+
+	value = "950823120"
+	ts_millis, _ = ConvertTimestampToMillis(value)
+	assert.Equal(t, uint64(950823120000), ts_millis)
+
+	value = "2019-06-11T16:33:51Z"
+	ts_millis, _ = ConvertTimestampToMillis(value)
+	assert.Equal(t, uint64(1560270831000), ts_millis)
+
+	value = "2020-08-03T07:10:20.123456+02:00"
+	ts_millis, _ = ConvertTimestampToMillis(value)
+	assert.Equal(t, uint64(1596431420123), ts_millis)
+
+	// nano seconds
+	value = "1714352490251000000"
+	ts_millis, _ = ConvertTimestampToMillis(value)
+	assert.Equal(t, uint64(1714352490251), ts_millis)
+
+	// invalid timestamps
+	value = "random string"
+	_, err := ConvertTimestampToMillis(value)
+	assert.NotNil(t, err)
+
+	value = "20201-08-03T07:10:20.123456+02:00"
+	_, err = ConvertTimestampToMillis(value)
+	assert.NotNil(t, err)
 }
